+++ generators/app/templates/grunt/config/notify_hooks.js
@@ -1,0 +1,19 @@
+/**
+ * Generated header by C0ZEN for generator-altran-angular project
+ * Generated file notify_hooks on WebStorm
+ *
+ * Created by: Geoffrey "C0ZEN" Testelin
+ * Date: 19/11/2017
+ * Time: 22:36
+ * Version: 1.0.0
+ */
+/* jscs:disable requireCamelCaseOrUpperCaseIdentifiers */
+module.exports = {
+	options: {
+		enabled                 : true,
+		max_jshint_notifications: 5,
+		success                 : false,
+		duration                : 5,
+		windowsVerbatimArguments: true
+	}
+};